+++ packages/graphiql-plugin-doc-explorer/__mocks__/monaco-editor.ts
@@ -1,0 +1,1 @@
+../../graphiql/__mocks__/monaco-editor.ts