+++ packages/graphiql-react/setup-files.ts
@@ -1,0 +1,4 @@
+// to make it works like Jest (auto-mocking)
+vi.mock('monaco-editor');
+
+export {};